todoist/pyproject.toml
[project]
name = "todoist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.3.0",
    "python-dotenv>=1.0.1",
    "todoist-api-python>=2.1.7",
]
